Add GitHub Packages publishing workflow
This commit adds a new workflow file `gh-registry-publish.yml` that sets up a job to publish packages to GitHub Packages. The workflow is triggered on release events of type "published". The job runs on the latest Ubuntu environment and has read access to repository contents and write access to packages. It performs the following steps:
- Checks out the repository using `actions/checkout@v4`
- Sets up Node.js version 18.x with the registry URL set to "https://npm.pkg.github.com" using `actions/setup-node@v3`
- Installs dependencies using `npm ci`
- Creates a unique version for the package if the branch is "main", by combining UTC timestamp and short SHA of the latest commit
- Builds the project using `npm run build`
- Publishes the package using `npm publish`, with authentication provided by `$GITHUB_TOKEN`